# schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "BookmarksViewGetEnvelope": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PilotBookmark"
            },
            "type": "array"
          },
          "meta": {
            "additionalProperties": {},
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
      "BookmarksViewPostEnvelope": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PilotBookmarkCreated"
          },
          "meta": {
            "additionalProperties": {},
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
      "BookmarksViewPostRequest": {
        "properties": {
          "artifact_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "artifact_id"
        ],
        "type": "object"
      },
...
      "IdentityClaimsViewGetEnvelope": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PilotIdentityClaim"
            },
            "type": "array"
          },
          "meta": {
            "additionalProperties": {},
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
      "IdentityClaimsViewPostEnvelope": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PilotIdentityClaim"
          },
          "meta": {
            "additionalProperties": {},
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "in": "cookie",
        "name": "sessionid",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Session-authenticated, permission-checked pilot API. Synthetic demo data is not real evidence.",
    "title": "Philanthra Pilot API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/pilot/bookmarks/": {
      "get": {
        "operationId": "pilot_Bookmarks_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Workspace-ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarksViewGetEnvelope"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "pilot"
        ]
      },
      "post": {
        "operationId": "pilot_Bookmarks_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Workspace-ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarksViewPostRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/BookmarksViewPostRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/BookmarksViewPostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarksViewPostEnvelope"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "pilot"
        ]
      }
    },
...
    "/api/v1/pilot/identity-claims/": {
      "get": {
        "operationId": "pilot_IdentityClaims_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Workspace-ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityClaimsViewGetEnvelope"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "pilot"
        ]
      },
      "post": {
        "operationId": "pilot_IdentityClaims_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Workspace-ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentityClaimsViewPostRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/IdentityClaimsViewPostRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/IdentityClaimsViewPostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityClaimsViewPostEnvelope"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "pilot"
        ]
      }
    },
    "/api/v1/pilot/identity-claims/{id}/proof/": {
      "get": {
        "operationId": "pilot_IdentityProof_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Workspace-ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityProofViewGetEnvelope"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "pilot"
        ]
      }
    },
...
    "/api/v1/pilot/introductions/{id}/contact/": {
      "get": {
        "operationId": "pilot_IntroductionContact_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Workspace-ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntroductionContactViewGetEnvelope"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "pilot"
        ]
      }
    },
...
    "/api/v1/pilot/password-reset/confirm/": {
      "post": {
        "operationId": "pilot_PasswordResetConfirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetConfirmViewPostRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetConfirmViewPostRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetConfirmViewPostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetConfirmViewPostEnvelope"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {}
        ],
        "tags": [
          "pilot"
        ]
      }
    },
    "/api/v1/pilot/password-reset/request/": {
      "post": {
        "operationId": "pilot_PasswordResetRequest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequestViewPostRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequestViewPostRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequestViewPostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetRequestViewPostEnvelope"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {}
        ],
        "tags": [
          "pilot"
        ]
      }
    },
    "/api/v1/pilot/profile/": {
      "get": {
        "operationId": "pilot_Profile_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Workspace-ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileViewGetEnvelope"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "pilot"
        ]
      },
      "patch": {
        "operationId": "pilot_Profile_patch",
        "parameters": [
          {
            "in": "header",
            "name": "X-Workspace-ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchedProfileViewPatchRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/PatchedProfileViewPatchRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/PatchedProfileViewPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileViewPatchEnvelope"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "pilot"
        ]
      }
    },
    "/api/v1/pilot/watchlist/": {
      "get": {
        "operationId": "pilot_Watchlist_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Workspace-ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistViewGetEnvelope"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "pilot"
        ]
      },
      "post": {
        "operationId": "pilot_Watchlist_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Workspace-ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistViewPostRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistViewPostRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistViewPostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistViewPostEnvelope"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "pilot"
        ]
      }
    },
    "/api/v1/pilot/watchlist/{id}/": {
      "delete": {
        "operationId": "pilot_WatchDetail_delete",
        "parameters": [
          {
            "in": "header",
            "name": "X-Workspace-ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No response body"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "pilot"
        ]
      }
    },
    "/api/v1/pilot/workspaces/": {
      "post": {
        "operationId": "pilot_Workspaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspacesViewPostRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/WorkspacesViewPostRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/WorkspacesViewPostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspacesViewPostEnvelope"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "pilot"
        ]
      }
    },
...
  }
}